blogservice/pyproject.toml
[tool.poetry]
name = "blogservice"
version = "0.1.0"
description = ""
authors = ["Egor Fironov <131647674+Egor1511@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
fastapi = {extras = ["all"], version = "^0.112.2"}
asyncpg = "^0.29.0"
requests = "^2.32.3"
pydantic-settings = "^2.4.0"
pycryptodome = "^3.20.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"
bcrypt = "4.0.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
isort = "^5.13.2"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pytest = "^8.3.2"

[tool.black]
line-length = 120
target-version = ['py39']
skip-string-normalization = false
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | _build
  | buck-out
  | build
  | dist
  | migration
)/
'''

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["your_project_name"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
